+++ CHANGELOG.md
@@ -13,6 +13,7 @@
   * new: node: access node API with bearer token as credentials
   * new: agent: `direct` allows ignoring certificate for wss using http options
   * new: preview: command `show` generates a preview and displays it in terminal
+  * new: global: option `fields` now supports `RegExp`
   * change(break): config: commands `detect` and `wizard` takes now a mandatory argument: address or url instead of option `url`.
   * change(break): sync: plugin `sync` is removed: actions are available through `server` and `node` plugins.
   * change(break): sync: replaced option `sync_session` with optional positional parameter.